requirements.txt
asyncio
ffmpeg-python
pyrogram
pytgcalls
tgcrypto
